mali400: reconfigure GPU (3)
The phone is consuming huge amount of power, thus frequency of 267MHz and above for GPU which uses more than 1A causes graphic stuttering, which should be an indication that the GPU is not running stable.
Lower step 1 (L0) frequency to 200MHz seems to help, at 0.95-0.975-1.0A. The massive stuttering is now unproducible, however graphic is still not as smooth as how it is on Android 9.