mali400: use 160MHz during FHD decoding
After thorough testing, we found that 160MHz is the most appropriate frequency for FHD decoding, and 133MHz for HD decoding.
However, mali400 tends to reside on 133MHz during FHD decoding, which causes minor stutters in playback.
Override frequency to 160MHz when FHD decoding is detected, if the targeted frequency is 133MHz.

CONFIG_VIDEO_MFC5X_DECODE_STATUS must be enabled for this to work.

Change-Id: Ic2fe5131c4422ecd63fd26a632b585ae80d195df